Changes in last problems plus another 3 problems

+++ Algebra/BiggestDivisorOf2Numbers/main.cpp
@@ -1,18 +1,18 @@
 #include <iostream>
 #include <vector>
 
-long long noodOfTwoNumbers(long long x, long long y);
+__int128 noodOfTwoNumbers(__int128 x, __int128 y);
 
 int main() {
-	long long x = 0;
-	long long y = 0;
+	__int128 x = 0;
+	__int128 y = 0;
 	std::cin >> x >> y;
-	long long ans = noodOfTwoNumbers(x,y);
+	__int128 ans = noodOfTwoNumbers(x,y);
 	std::cout << ans;
 	return 0;
 }
 
-long long noodOfTwoNumbers(long long x, long long y) {
+__int128 noodOfTwoNumbers(__int128 x, __int128 y) {
 	if (x == 0) {
 		return y;
 	} else {
